// src/app/@core/interfaces/auction/grantees.model.ts
export interface GranteesInterface {
    id?: number;
    description?: string;
    puesto?: string;
    type?: number;
    razonSocial?: string;
    street?: string;
    noInside?: string;
    noExterior?: string;
    col?: string;
    nommun?: string;
    nomedo?: string;
    cp?: string;
    usrStatus?: string;
}